add web

+++ src/main/java/com/suns/lottery/tball/bean/Dlt.java
@@ -36,9 +36,9 @@ public class Dlt {
 
     private Integer b2;
 
-    private Integer sales;
+    private Long sales;
 
-    private Integer poolmoney;
+    private Long poolmoney;
 
     private String content;
